feat: 프로필 정보, 권한 추가

+++ tangsquad/src/main/java/backend/tangsquad/domain/User.java
@@ -18,7 +18,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column private Integer id;
+    @Column private Long id;
 
     // 이름
     @Column(nullable = false) private String name;
